# docs/swagger.yaml
basePath: /api
definitions:
...
  server_internal_controller_http_V1_request.ForgotPassword:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - code
    - email
    - new_password
    type: object
  server_internal_controller_http_V1_request.Login:
    properties:
      captcha:
        type: string
      captcha_id:
        maxLength: 64
        minLength: 1
        type: string
      email:
        type: string
      password:
        maxLength: 20
        minLength: 1
        type: string
    required:
    - captcha
    - captcha_id
    - email
    - password
    type: object
  server_internal_controller_http_V1_request.Register:
    properties:
      code:
        type: string
      email:
        type: string
      password:
        maxLength: 20
        minLength: 8
        type: string
      username:
        maxLength: 32
        minLength: 2
        type: string
    required:
    - code
    - email
    - password
    - username
    type: object
...
  server_internal_controller_http_V1_response.Login:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
...
  server_internal_controller_http_V1_response.Refresh:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
  server_internal_controller_http_V1_response.Register:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
      user:
        $ref: '#/definitions/server_internal_controller_http_V1_response.UserProfile'
    type: object
...
  server_internal_controller_http_V1_response.UserProfile:
    properties:
      avatar:
        type: string
      bio:
        type: string
      blog_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      name:
        type: string
      region:
        type: string
      show_full_profile:
        type: boolean
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  server_internal_controller_http_shared.Envelope:
    properties:
      code:
        example: "0000"
        type: string
      data: {}
      message:
        example: ok
        type: string
    type: object
info:
  contact: {}
  description: Nlxiaoy Blog 后端 API（Admin + Public V1）。
  title: Nlxiaoy Blog API
  version: "1.0"
paths:
...
  /v1/auth/forgot:
    post:
      consumes:
      - application/json
      parameters:
      - description: 邮箱验证码与新密码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_controller_http_V1_request.ForgotPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      summary: 忘记密码
      tags:
      - V1.Auth
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: 登录成功后通过 HttpOnly Cookie 写入 refresh_token，同时响应体返回 access_token。
      parameters:
      - description: 登录信息（含图形验证码）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_controller_http_V1_request.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_V1_response.Login'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      summary: 用户登录
      tags:
      - V1.Auth
  /v1/auth/logout:
    post:
      description: 需要 BearerAuth（由路由层鉴权）；接口会清空 refresh_token Cookie。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      security:
      - BearerAuth: []
      summary: 退出登录
      tags:
      - V1.Auth
  /v1/auth/refresh:
    post:
      description: 读取 refresh_token Cookie，返回新的 access_token，并刷新 refresh_token Cookie。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_V1_response.Refresh'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      summary: 刷新令牌
      tags:
      - V1.Auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 注册信息（含邮箱验证码）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server_internal_controller_http_V1_request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/server_internal_controller_http_V1_response.Register'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      summary: 用户注册
      tags:
      - V1.Auth
...
  /v1/files/{object_key}:
    get:
      parameters:
      - description: 对象 Key
        in: path
        name: object_key
        required: true
        type: string
      responses:
        "307":
          description: Temporary Redirect
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
        "502":
          description: Bad Gateway
          schema:
            $ref: '#/definitions/server_internal_controller_http_shared.Envelope'
      summary: 获取文件临时访问 URL（重定向）
      tags:
      - V1.Files
...
schemes:
- http
- https
securityDefinitions:
  AdminSession:
    in: cookie
    name: fiber_session
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"